src/db_populate/requirements.txt
sqlalchemy
psycopg2-binary
bcrypt
faker
